Return errorObj in post('/upload') route

+++ src/backend/routes.js
@@ -1,6 +1,6 @@
 require('dotenv').load()
 const _ = require('lodash')
-const stringSimilarity = require('string-similarity')
+
 const express = require('express')
 const { ApolloServer } = require('apollo-server-express')
 
@@ -80,62 +80,40 @@ MongoClient.connect(process.env.LOADER_URI, { useNewUrlParser: true }, (err, cli
     res.send(createdCollection.collectionName)
   })
 
-  subApp.post('/upload', async (req, res, next) => {
-    const rawJson = req.body.data
+  const getErrorObj = require('./validation/getErrorObj')
 
-    if (rawJson[0].indication) { // TODO: really? does every doc have the same keys?
-      const indications = await pulseCoreDb.collection('indications').find().toArray()
-      const validIndications = _.keyBy(indications, 'name')
+  subApp.post('/upload', async ({
+    body: {
+      data,
+      collectionName,
+    }
+  }, res, next) => {
+    const errorObj = await getErrorObj(data,pulseCoreDb)
 
-      // add 1 for zero indexing, add 3 for rows skipped
-      // (there are two special rows in the xlsx file and the header)
-      const ROWS_TO_SKIP = 4
+    /*
+      ! Note on Error Management
+      * Currently just sending an error slice to the frontend to manually throw
+      * Still not sure how to accurately bubble up an express error to the frontend's catch
+      * Might be solved when error handling is moved to graphql
+    */
 
-      const problemRows = []
-      const invalidIndications = rawJson.filter(({ indication }, i) => {
-        const isIndicationInvalid = !validIndications[indication]
-        if (isIndicationInvalid) problemRows.push(i + ROWS_TO_SKIP)
-        return isIndicationInvalid
-      })
+    const hasErrors = !_.isEmpty(errorObj)
 
-      const numInvalidIndications = invalidIndications.length
-
-      if (numInvalidIndications > 0) {
-        const uniqueInvalidIndications = _.uniqBy(invalidIndications, 'indication').map(({ indication }) => indication)
-
-        const validIndicationArr = Object.keys(validIndications)
-        const suggestions = uniqueInvalidIndications.map(invalidIndication => {
-          const { bestMatch: { target } } = stringSimilarity.findBestMatch(invalidIndication, validIndicationArr)
-          return { 'Invalid Indication': invalidIndication, 'Did you mean...?': target }
-        })
-
-        let errorMessage = `
-          Indication validation failed!
-          Incoming data has ${numInvalidIndications} invalid indication entries.
-          Problem rows in CSV are: ${problemRows.join(', ')}
-          Your unique invalid indications are:
-          ${ suggestions }
-        `
-
-        console.table(suggestions, ['Invalid Indication', 'Did you mean...?'])
-
-        next(errorMessage)
-        return
-      }
+    if (hasErrors) {
+      res.status(400)
+      res.send({ error: errorObj })
+      return
     }
 
-    debugger
-
-    const targetCollection = pulseRawDb.collection(req.body.collectionName)
+    const targetCollection = pulseRawDb.collection(collectionName)
 
     await targetCollection.deleteMany()
-    await targetCollection.insertMany(req.body.data)
+    await targetCollection.insertMany(data)
 
     const persistedData = await targetCollection.find().toArray()
 
     res.json(persistedData)
   })
 })
-
 
 module.exports = subApp
